Most instructions now implemented
To clarify on the last commit because I forgot to add a better commit
message before I pushed it:
Created instructions.rs which consists mostly of an enum with variants
for every single instruction, and a way to transform a u16 value into
the matching instruction (if any)

This commit now adds on the instructions providing an actual
implementation of what they are supposed to do. Everything cpu related
or that didn't rely on either keyboard or display being ready to use was
implemented.